move gocty helpers to shared public directory
