+++ internal/app/decker/plugins/metasploit/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	// "encoding/json"
 	"fmt"
-	"github.com/stevenaldinger/decker/internal/pkg/gocty"
+	"github.com/stevenaldinger/decker/pkg/gocty"
 	"github.com/zclconf/go-cty/cty"
 	"os"
 	"os/exec"
