+++ internal/app/decker/plugins/wpscan/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/stevenaldinger/decker/internal/pkg/gocty"
+	"github.com/stevenaldinger/decker/pkg/gocty"
 	"github.com/zclconf/go-cty/cty"
 	"os"
 	"os/exec"
