+++ pkg/gocty/schema.go
